Move tests from test_init.py to test_switch.py
